==> analyses/01A-wrangle-specimen-data.Rmd <==
---
title: "Wrangle pangolin data"
output: html_document
---

This script takes georeferenced data for pangolins and wrangles these for further analysis.

## Load libraries
```{r, message = FALSE, warning = FALSE}
## Install sfe from GitHub
##devtools::install_github("JCur96/sfe", force = TRUE)

library(sfe)
library(rgeos)
library(rgdal)
library(Hmisc)
library(tidyverse)
library(here)
library(patchwork)
library(knitr)

# Short function to get decades
floor_decade <- function(x){
  if(!is.na(x)){
  x - x %% 10
  }else{NA_character_}
}
```

## Read in the georeferenced data
```{r, message = FALSE, warning = FALSE}
## NHM georeferenced data
nhm <- read_csv(here("raw-data/NHM_pangolins_June2019.csv"))

# Remove material column and unknown species names or port specimens
nhm <- 
  nhm %>% 
  select(-Material) %>%
  filter(Binomial != "unknown") %>%
  # Remove specimens from ports
  filter(RegistrationNumber != "NHMUK_12.12.3.3" & 
         RegistrationNumber != "NHMUK_no reg 9") %>%
  # Remove specimens from Sudan as this must be an error.
  filter(RegistrationNumber != "NHMUK_87.12.1.103" & 
         RegistrationNumber != "NHMUK_7.12.1.102") %>%
  # Remove specimens which may be errors 
  filter(RegistrationNumber != "NHMUK_27.12.1.203")
  


## GBIF georeferenced data
gbif <- read_csv(here("raw-data/gbif_pangolins_Oct2019.csv"))

## Exclude the specimens which are either misidentified or have incorrect localities
## These are Phataginus species found in southern Africa very far from their current range, or Asia where the genus is not found. Some of these could be Smutsia, but we can't check the specimens. See text for more details.
gbif <-
  gbif %>%
  filter(RegistrationNumber != "BOUM_11.Man.1" & 
         RegistrationNumber != "MVZ_MVZ:Mamm:4829" &
         RegistrationNumber != "AMNH_M-216259" &
         RegistrationNumber != "SUI_18440" &
         RegistrationNumber != "O_18156" &
         RegistrationNumber != "GNM_Ma ex 1292" &
         RegistrationNumber != "GNM_CollAn 7491" &
         RegistrationNumber != "MNHN_MO-1899-647" &
         RegistrationNumber != "UMMZ_156537" &
        RegistrationNumber != "MCZ_38282")
```

## Read in the ecology data
```{r, message = FALSE, warning = FALSE}
ecology <- read_csv(here("raw-data/pangolin-ecology.csv"))
```

## Add/Merge all the georeferenced and ecology data together
```{r}
specs <- 
  nhm %>%
  rbind(gbif) %>%
  full_join(ecology, by = c("Binomial" = "binomial"))
```

## Identify duplicates

- these are either specimens from the same species with the exact same latitude and longitude coordinates or...
- specimens from the same species with the exact same latitude and longitude coordinates and the same year
- adds two columns identifying duplicates to exclude in analyses

```{r}
specsx <-
  specs %>%
  dplyr::select(Binomial, Latitude, Longitude)

specsx_year <-
  specs %>%
  dplyr::select(Binomial, Latitude, Longitude, Year)

specs <- 
  specs %>%
  mutate(duplicates = duplicated(specsx)) %>%
  mutate(duplicates_year = duplicated(specsx_year))
```
## Tidy the georeferenced data

- Change species binomial columns to "binomial" as many of the functions rely on the data having a column of species names under the column header of "binomial". 
- Filter out blank or NA values from binomial, latitude and longitude. 
- Convert to the sf format, ensuring correct projection in latitude and longitude.
  
```{r}
## Tidy georeferenced data
specs <- 
  specs %>%
  rename(binomial = Binomial) %>%
  # Exclude specimens with no name or coordinates
  filter(!is.na(binomial) & binomial != "unknown" &
         !is.na(Latitude) & !is.na(Longitude)) %>%
  # Add decade variable (function above)
  # This maps to character to deal with NAs so needs coercing back to numeric
  mutate(Decade = map_chr(Year, floor_decade)) %>%
  mutate(Decade = as.numeric(Decade)) %>%
  # Convert extents to km
  mutate(Extent_km = Extent_m / 1000) %>%
  # Remove non needed columns
  dplyr::select(-NOTES, -LocalityType, -Locality, -Extent_m) %>%
  # Convert to sf format
  st_as_sf(coords = c("Longitude", "Latitude"), crs = 4326)
```

## Add error polygons to data

- change the Extent column into kilometers and rename it
- add a `buffer` to the point data using the point-radius method.
```{r, message = FALSE, warning = FALSE}
specs_errors <-
  specs %>%
  addError()
```

## Save the outputs
Note that `st_write` will throw an error if you try to overwrite a file, so delete previous version of the files first. To avoid this error I've commented these out here.
```{r}
#st_write(specs, here("data/specimens-points.csv"), 
         #layer_options = "GEOMETRY=AS_WKT")
#st_write(specs_errors, here("data/specimens-extents.csv"), 
         #layer_options = "GEOMETRY=AS_WKT")
```
## Summary info

```{r}
specs %>%
  # Remove specimens with low certainty and large extent
  filter(Certainty >= 50 & Extent_km < 50) %>%
  group_by(binomial) %>%
  summarise(n())
```

```{r}
specs %>%
  # Remove specimens with low certainty and large extent
  filter(Certainty >= 50 & Extent_km < 50) %>%
  # Remove duplicates
  filter(duplicates == 0) %>%
  group_by(binomial) %>%
  summarise(n())

```

## By years....
Remember one year is NA so you need to minus 1 from total numbers
```{r}
specs %>%
  # Remove specimens with low certainty and large extent
  filter(Certainty >= 50 & Extent_km < 50) %>%
  group_by(binomial, Year) %>%
  summarise(n()) %>%
  add_count(binomial,Year) %>%
  summarise(sum(n))
```


```{r}
specs %>%
  # Remove specimens with low certainty and large extent
  filter(Certainty >= 50 & Extent_km < 50) %>%
  # Remove duplicates
  filter(duplicates == FALSE) %>%
  group_by(binomial, Year) %>%
  summarise(n()) %>%
  add_count(binomial,Year) %>%
  summarise(sum(n))
```
## Totals

```{r}
specs %>%
  summarise(n())
```


```{r}
specs %>%
  # Remove specimens with low certainty and large extent
  filter(Certainty >= 50 & Extent_km < 50) %>%
  summarise(n())
```

```{r}
specs %>%
  # Remove specimens with low certainty and large extent
  filter(Certainty >= 50 & Extent_km < 50) %>%
  # Remove duplicates
  filter(duplicates == FALSE) %>%
  summarise(n())
```
