==> analyses/01C-wrangle-overlaps-outputs.Rmd <==
---
title: "Extract overlaps dataset"
output: html_document
---

```{r, message = FALSE}
library(tidyverse)
library(here)
```
Read in the specimen dataset and remove geometry
```{r, message = FALSE}
specs <- read_csv(here("data/specimens-extents.csv"))

specs <- 
  specs %>%
  select(-WKT)
```
Read in the AOH data and remove defunct columns and _ from species names
```{r, message = FALSE}
files <- paste0("raw-data/AOH_Overlaps/", 
                dir(path = here("raw-data/AOH_Overlaps"), 
                    pattern = "*.csv"))
  
aoh <- 
  files %>%
  here() %>%
  map(read_csv) %>%    
  reduce(rbind) %>%
  select(-percentOverlap, -X_id) %>%
  mutate(binomial = str_replace(binomial, "_", " "))

# Look at it
# aoh
```

Merge with the specimen data, but remove specimens without AOH overlaps, i.e. Smutsia temmincki.
```{r}
overlaps_aoh <- 
  aoh %>%
  left_join(specs, by = c("RegistrationNumber", "binomial")) %>%
  distinct()
```

Read in the IUCN data and remove defunct columns and _ from species names
```{r, message = FALSE}
files <- paste0("raw-data/IUCN_Overlaps/", 
                dir(path = here("raw-data/IUCN_Overlaps"), 
                    pattern = "*.csv"))
  
iucn <- 
  files %>%
  here() %>%
  map(read_csv) %>%    
  reduce(rbind) %>%
  select(-percentOverlap, -X_id) %>%
  mutate(binomial = str_replace(binomial, "_", " "))

# Look at it
# iucn
```

Merge with the specimen data
```{r}
overlaps_iucn <- 
  specs %>%
  full_join(iucn, by = c("RegistrationNumber", "binomial")) %>%
  distinct()
```

Create dataset with AOH for all but temminicki
```{r}
temm <- 
  overlaps_iucn %>%
  filter(binomial == "Smutsia temminckii")

overlaps_combined <- rbind(overlaps_aoh, temm)

# Look at it
# overlaps_combined
```

Save the files for use in models
```{r}
#write_csv(overlaps_iucn, file = here("data/overlaps_iucn.csv"))
write_csv(overlaps_aoh, path = here("data/overlaps_aoh.csv"))
write_csv(overlaps_combined, path = here("data/overlaps_combined.csv"))
```
