==> analyses/03-overlaps-analyses.Rmd <==
---
title: "Overlaps analyses 1 - extract summary data"
output: html_document
---
  
This script takes georeferenced data and IUCN range maps for pangolins and analyses the number of overlaps in various ways.

## Load libraries
```{r, message = FALSE, warning = FALSE}
library(sf)
library(sfe)
library(tidyverse)
library(here)
library(broom)
library(ggfortify)
library(knitr)

# SE function
se <- 
  function(y) {
  sqrt(var(y)/length(y))
}
```

## Prepare for the analyses
First read in the data as CSV not spatial data as we don't need the spatial geometries here. Relevel redlist so it matches increasing extinction risk.
```{r, message = FALSE, warning = FALSE}
overlaps <- 
  read_csv(here("data/overlaps_combined.csv")) %>%  
  mutate(redlist = factor(redlist, levels = c("VU", "EN", "CR")))
```

We now create four data subsets. 1 is all the data (except zero certainty specimens), 2 is all the data (except zero certainty specimens) minus duplicates, 3 is the data minus those with large extents and low certainty, 4 is the data minus those with large extents and low certainty, and with no duplicates. It needs to be done like this so we can properly count up the number of range overlapping and non range overlapping specimens.

### All data
```{r}
overlaps_all <- 
  overlaps %>%
  filter(Certainty > 0) %>%
  # Number of specimens per species total
  countNumbSpecimens() %>%
  # Number of overlaps per species
  countNumbOverlaps()
```

### Remove duplicates
```{r}
overlaps_all_nodups <- 
  overlaps %>%
  filter(Certainty > 0) %>%
  filter(duplicates == 0) %>%
  # Number of specimens per species total
  countNumbSpecimens() %>%
  # Number of overlaps per species
  countNumbOverlaps()
```

### Omit specimens with large extents and low certainty
```{r}
overlaps_good <-
  overlaps %>%
  filter(Certainty >= 50 & Extent_km < 50) %>%
  # Number of specimens per species total
  countNumbSpecimens() %>%
  # Number of overlaps per species
  countNumbOverlaps()
```

### Omit specimens with large extents and low certainty and duplicates
```{r}
overlaps_good_nodups <-
  overlaps %>%
  filter(Certainty >= 50 & Extent_km < 50) %>%
  filter(duplicates == 0) %>%
  # Number of specimens per species total
  countNumbSpecimens() %>%
  # Number of overlaps per species
  countNumbOverlaps()
```

## Summarising the degree of overlap
This is for Table S3

All data

### Species

```{r}
all_species <-
overlaps_all %>%
  group_by(binomial) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_species)
```

### Continent

```{r}
all_continents <-
overlaps_all %>%
  group_by(Continent) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_continents)
```

### Ecology

```{r}
all_ecology <-
overlaps_all %>%
  group_by(ecology) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_ecology)
```

### IUCN Red List

```{r}
all_iucn <-
overlaps_all %>%
  group_by(redlist) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_iucn)
```
All data minus duplicates

### Species

```{r}
all_species_nodups <-
overlaps_all_nodups %>%
  group_by(binomial) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_species_nodups)
```

### Continent

```{r}
all_continents_nodups <-
overlaps_all_nodups %>%
  group_by(Continent) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_continents_nodups)
```

### Ecology

```{r}
all_ecology_nodups <-
overlaps_all_nodups %>%
  group_by(ecology) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_ecology_nodups)
```

### IUCN Red List

```{r}
all_iucn_nodups <-
overlaps_all_nodups %>%
  group_by(redlist) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(all_iucn_nodups)
```

"Good"" data only

### Species

```{r}
good_species <-
overlaps_good %>%
  group_by(binomial) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_species)
```

### Continent

```{r}
good_continents <-
overlaps_good %>%
  group_by(Continent) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_continents)
```

### Ecology

```{r}
good_ecology <-
overlaps_good %>%
  group_by(ecology) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_ecology)
```

### IUCN Red List

```{r}
good_iucn <-
overlaps_good %>%
  group_by(redlist) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_iucn)
```

"Good" data minus duplicates

### Species

```{r}
good_species_nodups <-
overlaps_good_nodups %>%
  group_by(binomial) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_species_nodups)
```

### Continent

```{r}
good_continents_nodups <-
overlaps_good_nodups %>%
  group_by(Continent) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_continents_nodups)
```

### Ecology

```{r}
good_ecology_nodups <-
overlaps_good_nodups %>%
  group_by(ecology) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_ecology_nodups)
```

### IUCN Red List

```{r}
good_iucn_nodups <-
overlaps_good_nodups %>%
  group_by(redlist) %>%
  summarise(n(),
            overlap = mean(numberOfOverlaps)/mean(numberOfSpecimens) * 100,
            mnpercent = mean(Percent_overlap),
            sepercent = se(Percent_overlap)) %>%
  mutate_if(is.numeric, round, digits = 2)

kable(good_iucn_nodups)
```
