==> analyses/04-overlaps-analyses-models.Rmd <==
---
title: "Overlaps analyses 2 - models"
output: html_document
---
  
This script takes georeferenced data and IUCN range maps for pangolins and analyses the number of overlaps in various ways.

## Load libraries
```{r, message = FALSE, warning = FALSE}
library(sf)
library(sfe)
library(tidyverse)
library(here)
library(broom)
library(ggfortify)
library(betareg)
library(lmtest)
```

## Prepare for the analyses
First read in the data as CSV not spatial data as we don't need the spatial geometries here. Relevel redlist so it matches increasing extinction risk. Omit specimens with large extents and low certainty. Omit duplicates and add success failure numbers.
```{r, message = FALSE, warning = FALSE}
overlaps_all <- 
  read_csv(here("data/overlaps_combined.csv")) %>%  
  mutate(redlist = factor(redlist, levels = c("VU", "EN", "CR"))) %>%
  filter(Certainty >= 50 & Extent_km < 50) %>%
  filter(duplicates == 0) %>%
  # Number of specimens per species total
  countNumbSpecimens() %>%
  # Number of overlaps per species
  countNumbOverlaps()
```

Repeat but for years data which has different duplicates:
```{r}
overlaps_year <- 
  read_csv(here("data/overlaps_combined.csv")) %>%  
  mutate(redlist = factor(redlist, levels = c("VU", "EN", "CR"))) %>%
  filter(Certainty >= 50 & Extent_km < 50) %>%
  filter(duplicates_year == 0) %>%
  # Number of specimens per species total
  countNumbSpecimens() %>%
  # Number of overlaps per species
  countNumbOverlaps()
```

## Models with number of overlaps
These are binomial GLMs but we only have 8 species so too little power to say much. Included mainly for completeness.

First create new datasets only with unique values to avoid accidental pseudoreplication.
```{r}
overlaps_year_specs <-
  overlaps_year %>%
  dplyr::select(binomial, Year, numberOfOverlaps, numberOfSpecimens) %>%
  distinct()

overlaps_all_specs <-
  overlaps_all %>%
  dplyr::select(binomial, Continent, ecology, redlist, numberOfOverlaps, numberOfSpecimens) %>%
  distinct()
```

### Years
```{r}
model1a <- glm(cbind(numberOfOverlaps, numberOfSpecimens - numberOfOverlaps) ~ Year,   
               data = overlaps_year_specs, family = binomial)

## Model diagnostics
autoplot(model1a)

## Model outputs
anova(model1a, test = "Chisq")
```

```{r}
summary(model1a)
```

### Continent
```{r}
model1b <- glm(cbind(numberOfOverlaps, numberOfSpecimens - numberOfOverlaps) ~ Continent,   
               data = overlaps_all_specs, family = binomial)

## Model diagnostics
autoplot(model1b)
```

```{r}
## Model outputs
anova(model1b, test = "Chisq")
```

```{r}
summary(model1b)
```

### Ecology
```{r}
model1c <- glm(cbind(numberOfOverlaps, numberOfSpecimens - numberOfOverlaps) ~ ecology,   
               data = overlaps_all_specs, family = binomial)

## Model diagnostics
autoplot(model1c)
```

```{r}
## Model outputs
anova(model1c, test = "Chisq")
```

```{r}
summary(model1c)
```

### IUCN status
```{r}
model1d <- glm(cbind(numberOfOverlaps, numberOfSpecimens - numberOfOverlaps) ~ redlist,   
               data = overlaps_all_specs, family = binomial)

## Model diagnostics
autoplot(model1d)
```

```{r}
## Model outputs
anova(model1d, test = "Chisq")
```

```{r}
summary(model1d)
```

## Models - with % area overlaps
Beta regression with logit link

First rescale percent to 0-1
```{r}
overlaps_year$Percent_overlap <- overlaps_year$Percent_overlap / 100
overlaps_all$Percent_overlap <- overlaps_all$Percent_overlap / 100
```

Next rescale to remove zeros and 1s
```{r}
transform01 <- function(x) {
  (x * (length(x) - 1) + 0.5) / (length(x))
}

overlaps_year$Percent_overlap_scaled <- transform01(overlaps_year$Percent_overlap)
overlaps_all$Percent_overlap_scaled <- transform01(overlaps_all$Percent_overlap)
```

### Year
```{r}
# Fit model with fixed phi
model2a1 <- betareg(Percent_overlap_scaled ~ Year, data = overlaps_year)
# Fit model with variable phi
model2a2 <- betareg(Percent_overlap_scaled ~ Year | Year, data = overlaps_year)

# Which is better fit?
lrtest(model2a1, model2a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model2a1.bc <- betareg(Percent_overlap_scaled ~ Year, data = overlaps_year, type = "BC")
# Fit model with variable phi
model2a2.bc <- betareg(Percent_overlap_scaled ~ Year | Year, data = overlaps_year, type = "BC")
# Which is better fit?
lrtest(model2a1.bc, model2a2.bc)
```

```{r}
# Fit model with fixed phi
model2a1.br <- betareg(Percent_overlap_scaled ~ Year, data = overlaps_year, type = "BR")
# Fit model with variable phi
model2a2.br <- betareg(Percent_overlap_scaled ~ Year | Year, data = overlaps_year, type = "BR")

# Which is better fit?
lrtest(model2a1.br, model2a2.br)
```

```{r}
## Model outputs
summary(model2a1)
```

### Continent

```{r}
# Fit model with fixed phi
model2b1 <- betareg(Percent_overlap_scaled ~ Continent, data = overlaps_all)
# Fit model with variable phi
model2b2 <- betareg(Percent_overlap_scaled ~ Continent | Continent, data = overlaps_all)

# Which is better fit?
lrtest(model2b1, model2b2)
```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model2b1.bc <- betareg(Percent_overlap_scaled ~ Continent, data = overlaps_all, type = "BC")
# Fit model with variable phi
model2b2.bc <- betareg(Percent_overlap_scaled ~ Continent | Continent, data = overlaps_all, type = "BC")
# Which is better fit?
lrtest(model2b1.bc, model2b2.bc)
```

```{r}
# Fit model with fixed phi
model2b1.br <- betareg(Percent_overlap_scaled ~ Continent, data = overlaps_all, type = "BR")
# Fit model with variable phi
model2b2.br <- betareg(Percent_overlap_scaled ~ Continent | Continent, data = overlaps_all, type = "BR")

# Which is better fit?
lrtest(model2b1.br, model2b2.br)
```
```{r}
## Model outputs
summary(model2b2)
```

### Ecology
```{r}
# Fit model with fixed phi
model2c1 <- betareg(Percent_overlap_scaled ~ ecology, data = overlaps_all)
# Fit model with variable phi
model2c2 <- betareg(Percent_overlap_scaled ~ ecology | ecology, data = overlaps_all)

# Which is better fit?
lrtest(model2c1, model2c2)
```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model2c1.bc <- betareg(Percent_overlap_scaled ~ ecology, data = overlaps_all, type = "BC")
# Fit model with variable phi
model2c2.bc <- betareg(Percent_overlap_scaled ~ ecology | ecology, data = overlaps_all, type = "BC")
# Which is better fit?
lrtest(model2c1.bc, model2c2.bc)
```

```{r}
# Fit model with fixed phi
model2c1.br <- betareg(Percent_overlap_scaled ~ ecology, data = overlaps_all, type = "BR")
# Fit model with variable phi
model2c2.br <- betareg(Percent_overlap_scaled ~ ecology | ecology, data = overlaps_all, type = "BR")

# Which is better fit?
lrtest(model2c1.br, model2c2.br)
```
```{r}
## Model outputs
summary(model2c2)
```

### IUCN
```{r}
# Fit model with fixed phi
model2d1 <- betareg(Percent_overlap_scaled ~ redlist, data = overlaps_all)
# Fit model with variable phi
model2d2 <- betareg(Percent_overlap_scaled ~ redlist | redlist, data = overlaps_all)

# Which is better fit?
lrtest(model2d1, model2d2)
```

Check results are the same with bias correction and bias reduction
```{r}
# Fit model with fixed phi
model2d1.bc <- betareg(Percent_overlap_scaled ~ redlist, data = overlaps_all, type = "BC")
# Fit model with variable phi
model2d2.bc <- betareg(Percent_overlap_scaled ~ redlist | redlist, data = overlaps_all, type = "BC")
# Which is better fit?
lrtest(model2d1.bc, model2d2.bc)
```

```{r}
# Fit model with fixed phi
model2d1.br <- betareg(Percent_overlap_scaled ~ redlist, data = overlaps_all, type = "BR")
# Fit model with variable phi
model2d2.br <- betareg(Percent_overlap_scaled ~ redlist | redlist, data = overlaps_all, type = "BR")

# Which is better fit?
lrtest(model2d1.br, model2d2.br)
```

```{r}
## Model outputs
summary(model2d2)
```

### Species
```{r}
# Fit model with fixed phi
model2e1 <- betareg(Percent_overlap_scaled ~ binomial, data = overlaps_all)
# Fit model with variable phi
model2e2 <- betareg(Percent_overlap_scaled ~ binomial | binomial, data = overlaps_all)

# Which is better fit?
lrtest(model2e1, model2e2)
```

Check results are the same with bias correction and bias reduction
```{r}
# Fit model with fixed phi
model2e1.bc <- betareg(Percent_overlap_scaled ~ binomial, data = overlaps_all, type = "BC")
# Fit model with variable phi
model2e2.bc <- betareg(Percent_overlap_scaled ~ binomial | binomial, data = overlaps_all, type = "BC")
# Which is better fit?
lrtest(model2e1.bc, model2e2.bc)
```

```{r}
# Fit model with fixed phi
model2e1.br <- betareg(Percent_overlap_scaled ~ binomial, data = overlaps_all, type = "BR")
# Fit model with variable phi
model2e2.br <- betareg(Percent_overlap_scaled ~ binomial | binomial, data = overlaps_all, type = "BR")

# Which is better fit?
lrtest(model2e1.br, model2e2.br)
```

```{r}
## Model outputs
summary(model2e1)
```
