==> analyses/06-extract-land-use-data.Rmd <==
---
title: "Extract Land use change data"
output: html_document
---
  
This script takes extracts various types of land use data from the error polygons of the specimens.

## Load libraries
```{r, message = FALSE, warning = FALSE}
library(sf)
library(sfe)
library(tidyverse)
library(here)
library(ncdf4)
library(raster) 
library(rgdal)

source(here::here("analyses/functions-anthropocene-data.R"))
```

Read in the data and set coordinate ref system. For some reason st_read makes everything a factor so convert Extent etc back to numeric first and relevel redlist
```{r}
specs_errors <- 
  st_read(here("data/specimens-extents.csv")) %>%  
  st_set_crs(4326) %>%
  mutate_at(vars(Extent_km, Certainty), 
            ~as.numeric(as.character(.))) %>%
  # Omit poor accuracy specimens
  filter(Certainty >= 50 & Extent_km < 50) %>%
  # Omit duplicates 
  filter(duplicates == 0)
```

## Extracting land use data

We will use the LUH2 dataset for landuse, here's a list of the variables:

* primf: forested primary land
* primn: non-forested primary land
* urban: urban land

First we read in the data which is netCDF data...

```{r}
# Read in the netCDF data
nc_data <- nc_open(here("raw-data/spatial-data/LUH2_high_states.nc"))

# Capture the lat, lon and time variables
lon <- ncvar_get(nc_data, "lon")
lat <- ncvar_get(nc_data, "lat", verbose = F)
time <- ncvar_get(nc_data, "time")
```

Next we extract changes in land use variables within each point-polygon area for each specimen in the `overlaps` data. The differences are added to the data as new variables. Differences are also for various time periods.

The functions below will throw warnings about coordinate systems but all are in WGS84 so it's fine.

*This takes 5-10 minutes to run - go and get a coffee*

```{r, message = FALSE, warning = FALSE}
overlaps_landuse <- specs_errors

# Primary forest
overlaps_landuse <- 
  get_landuse_diff("primf", nc_data, overlaps_landuse) 

# Primary non forest
overlaps_landuse <- 
  get_landuse_diff("primn", nc_data, overlaps_landuse) 

# Urban
overlaps_landuse <- 
  get_landuse_diff("urban", nc_data, overlaps_landuse) 

# Close data set 
nc_close(nc_data)

# Write to file
st_write(overlaps_landuse, here("data/overlaps_landuse.csv"))
```
