==> analyses/07-land-use-analyses.Rmd <==
---
title: "Landuse change analyses"
output: html_document
---

## Load libraries
```{r, message = FALSE, warning = FALSE}
library(sf)
library(sfe)
library(tidyverse)
library(here)
library(broom)
library(ggfortify)
library(betareg)
library(lmtest)
```

## Prepare for the analyses
First read in the data as CSV not spatial data as we don't need the spatial geometries here. Relevel redlist so it matches increasing extinction risk.
```{r, message = FALSE, warning = FALSE}
overlaps_landuse <- 
  read_csv(here("data/overlaps_landuse.csv")) %>%  
  mutate(redlist = factor(redlist, levels = c("VU", "EN", "CR")))
```

Next read in overlaps data so we can get the number of overlaps
```{r}
overlaps <- 
  read_csv(here("data/overlaps_combined.csv")) 

overlaps <- 
  as.data.frame(overlaps) %>%
  dplyr::select(RegistrationNumber, Percent_overlap)

overlaps_landuse <- left_join(overlaps_landuse, overlaps, 
                              by = "RegistrationNumber")
```

## Models - with % area overlaps
Beta regression with logit link

First rescale percent to 0-1
```{r}
overlaps_landuse$Percent_overlap <- overlaps_landuse$Percent_overlap / 100
```

Next rescale to remove zeros and 1s
```{r}
transform01 <- function(x) {
  (x * (length(x) - 1) + 0.5) / (length(x))
}

overlaps_landuse$Percent_overlap_scaled <- 
  transform01(overlaps_landuse$Percent_overlap)
```

Variables are:
* primf: forested primary land
* primn: non-forested primary land
* urban: urban land

## Primary Forest
### primf_1850
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ primf_1850, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ primf_1850 | primf_1850, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ primf_1850, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ primf_1850 | primf_1850, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ primf_1850, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ primf_1850 | primf_1850, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a1)
```

### primf_1900
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ primf_1900, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ primf_1900 | primf_1900, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ primf_1900, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ primf_1900 | primf_1900, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ primf_1900, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ primf_1900 | primf_1900, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a1)
```

### primf_1950
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ primf_1950, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ primf_1950 | primf_1950, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ primf_1950, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ primf_1950 | primf_1950, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ primf_1950, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ primf_1950 | primf_1950, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a1)
```

## Non forested primary land
### primn_1850
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ primn_1850, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ primn_1850 | primn_1850, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ primn_1850, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ primn_1850 | primn_1850, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ primn_1850, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ primn_1850 | primn_1850, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a1)
```

### primn_1900
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ primn_1900, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ primn_1900 | primn_1900, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ primn_1900, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ primn_1900 | primn_1900, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ primn_1900, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ primn_1900 | primn_1900, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a1)
```

### primn_1950
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ primn_1950, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ primn_1950 | primn_1950, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ primn_1950, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ primn_1950 | primn_1950, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ primn_1950, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ primn_1950 | primn_1950, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a2)
```

# Urban
### urban_1850
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ urban_1850, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ urban_1850 | urban_1850, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ urban_1850, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ urban_1850 | urban_1850, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ urban_1850, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ urban_1850 | urban_1850, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a2)
```

### urban_1900
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ urban_1900, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ urban_1900 | urban_1900, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ urban_1900, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ urban_1900 | urban_1900, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ urban_1900, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ urban_1900 | urban_1900, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a2)
```

### urban_1950
```{r}
# Fit model with fixed phi
model1a1 <- betareg(Percent_overlap_scaled ~ urban_1950, data = overlaps_landuse)
# Fit model with variable phi
model1a2 <- betareg(Percent_overlap_scaled ~ urban_1950 | urban_1950, data = overlaps_landuse)

# Which is better fit?
lrtest(model1a1, model1a2)

```

Check results are the same with bias correction and bia reduction
```{r}
# Fit model with fixed phi
model1a1.bc <- betareg(Percent_overlap_scaled ~ urban_1950, data = overlaps_landuse, type = "BC")
# Fit model with variable phi
model1a2.bc <- betareg(Percent_overlap_scaled ~ urban_1950 | urban_1950, data = overlaps_landuse, type = "BC")
# Which is better fit?
lrtest(model1a1.bc, model1a2.bc)
```

```{r}
# Fit model with fixed phi
model1a1.br <- betareg(Percent_overlap_scaled ~ urban_1950, data = overlaps_landuse, type = "BR")
# Fit model with variable phi
model1a2.br <- betareg(Percent_overlap_scaled ~ urban_1950 | urban_1950, data = overlaps_landuse, type = "BR")

# Which is better fit?
lrtest(model1a1.br, model1a2.br)
```

```{r}
## Model outputs
summary(model1a2)
```
